Refactor cli/config into cli/credentials
