+++ internal/cli/credentials/credentials_test.go
@@ -1,4 +1,4 @@
-package config
+package credentials
 
 import (
 	"os"
@@ -49,7 +49,7 @@ func TestProviderFromJSON_ProvideForRegion(t *testing.T) {
 			t.Fatalf("expected no error, got %v", err)
 		}
 
-		expected := &OAuth{
+		expected := &OAuthCredentials{
 			ClientID:     "your-client-id-eu10",
 			ClientSecret: "your-client-secret-eu10",
 			TokenURL:     "https://oauth.cf-eu10-canary.test.com/oauth/token",
@@ -72,7 +72,7 @@ func TestProviderFromJSON_ProvideForRegion(t *testing.T) {
 			t.Fatalf("expected no error, got %v", err)
 		}
 
-		expected := &OAuth{
+		expected := &OAuthCredentials{
 			ClientID:     "your-client-id-us31",
 			ClientSecret: "your-client-secret-us31",
 			TokenURL:     "https://oauth.cf-us31.test.com/oauth/token",
@@ -151,7 +151,7 @@ func TestProviderFromJSON_ProvideForRegion(t *testing.T) {
 
 func TestValidateRegionCredentials(t *testing.T) {
 	t.Run("valid credentials", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientID:     "test-id",
 			ClientSecret: "test-secret",
 			TokenURL:     "https://oauth.example.com/token",
@@ -165,7 +165,7 @@ func TestValidateRegionCredentials(t *testing.T) {
 	})
 
 	t.Run("missing client_id", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientSecret: "test-secret",
 			TokenURL:     "https://oauth.example.com/token",
 			ServiceURL:   "https://api.example.com",
@@ -178,7 +178,7 @@ func TestValidateRegionCredentials(t *testing.T) {
 	})
 
 	t.Run("missing client_secret", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientID:   "test-id",
 			TokenURL:   "https://oauth.example.com/token",
 			ServiceURL: "https://api.example.com",
@@ -191,7 +191,7 @@ func TestValidateRegionCredentials(t *testing.T) {
 	})
 
 	t.Run("missing token_url", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientID:     "test-id",
 			ClientSecret: "test-secret",
 			ServiceURL:   "https://api.example.com",
@@ -204,7 +204,7 @@ func TestValidateRegionCredentials(t *testing.T) {
 	})
 
 	t.Run("missing service_url", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientID:     "test-id",
 			ClientSecret: "test-secret",
 			TokenURL:     "https://oauth.example.com/token",
@@ -217,7 +217,7 @@ func TestValidateRegionCredentials(t *testing.T) {
 	})
 
 	t.Run("multiple missing fields", func(t *testing.T) {
-		oauth := &OAuth{
+		oauth := &OAuthCredentials{
 			ClientID: "test-id",
 		}
